Fix bugs, rename files, but still have problems.
I made `_div_by_hand` because I want to avoid using `int(a/b)`,
but it's very slow for large `a`. (I am trying to fix this right now)

Why not using `a//b`? Because I remembered I got unexpected results in
`_round_int`, so then I tried to use `//` as few as possible.